SECONDS to HOUR ,MIN ,SEC
A simple rule to estimate your ideal body weight is to allow 110 pounds for the first 5 feet of height and 5 pounds for each additional inch. Write a program with a variable for the height of a person in feet and another variable for the additional inches. Assume the person is at least 5 feet tall Take Input From user for height as height in feet In separate variable and additional inches in separate variable.