count caffine in glass
Scientist estimates that roughly 10 grams cf caffeine consumed at one time is a lethal overdose. Write a program with a variable that holds the number of milligrams of caffeine in a drink and outputs how many drinks it takes to kill a person. A 12-ounce can of has approximate/ 34 mg of caffeine, while the Is-ounce cup of coffee has a pprox[mately ISO mg of caffeine.