time calculation by speed and distance
John travels 55 mi 25 at an average speed of IS mile5 per hour. Write a program to calculöte the total number of hours John takes to cover this distance. The program should print the total time taken in hours and minutes_ Use the following formula for calculations.
Distance
Time = 
Speed