grade for N classroom
Write a program that calculates the total grade for N classroom exercises as a percentage. Use the Decimal
Format class to output the value as a percent. The user should input the value for N followed by each of
the N scores and totals. Calculate the overall percentage (sum of the total points earned divided by the
total points possible) and output it using the Decimal Format class