four-digit Arabic
Write a program that accepts a year written as a four-digit Arabic (ordinary) numeral and outputs the year
written in roman numerals. Important roman numerals are V for 5, X for 10, L for 50, C for 100, d for 500,
and M for 1,000. recall that some numbers are formed by using a kind of subtraction of one roman “digit”;
for example, IV is 4 produced as V minus i, XL is 40, CM is 900, and so on. A few sample years: MCM is
1900, MCML is 1950, MCMLX is 1960, MCMXL is 1940, MCMLXXXiX is 1989. Assume the year is between
1000 and 3000. Your program should include a loop that lets the user repeat this calculation until the user
says she or he is done.